+++ wcfsetup/install/files/js/WCF.js
@@ -6777,11 +6777,16 @@ WCF.Sortable.List = Class.extend({
 			toleranceElement: '> span'
 		}, options || { });
 		
+		var sortableList = $('#' + this._containerID + ' .sortableList');
+		if (sortableList.is('tbody') && this._options.helper === 'clone') {
+			this._options.helper = this._tableRowHelper.bind(this);
+		}
+		
 		if (isSimpleSorting) {
-			$('#' + this._containerID + ' .sortableList').sortable(this._options);
+			sortableList.sortable(this._options);
 		}
 		else {
-			$('#' + this._containerID + ' > .sortableList').nestedSortable(this._options);
+			sortableList.nestedSortable(this._options);
 		}
 		
 		if (this._className) {
@@ -6796,6 +6801,23 @@ WCF.Sortable.List = Class.extend({
 			
 			$formSubmit.children('button[data-type="submit"]').click($.proxy(this._submit, this));
 		}
+	},
+	
+	/**
+	 * Fixes the width of the cells of the dragged table row.
+	 * 
+	 * @param	{Event}		event
+	 * @param	{jQuery}	ui
+	 * @return	{jQuery}
+	 */
+	_tableRowHelper: function(event, ui) {
+		ui.children('td').each(function(index, element) {
+			element = $(element);
+			
+			element.width(element.width());
+		});
+		
+		return ui;
 	},
 	
 	/**