+++ wcfsetup/install/files/lib/data/label/LabelList.class.php
@@ -6,7 +6,7 @@ use wcf\data\DatabaseObjectList;
  * Represents a list of labels.
  * 
  * @author	Alexander Ebert
- * @copyright	2001-2015 WoltLab GmbH
+ * @copyright	2001-2016 WoltLab GmbH
  * @license	GNU Lesser General Public License <http://opensource.org/licenses/lgpl-license.php>
  * @package	com.woltlab.wcf
  * @subpackage	data.label
@@ -14,7 +14,12 @@ use wcf\data\DatabaseObjectList;
  */
 class LabelList extends DatabaseObjectList {
 	/**
-	 * @see	\wcf\data\DatabaseObjectList::$className
+	 * @inheritDoc
 	 */
-	public $className = 'wcf\data\label\Label';
+	public $className = Label::class;
+	
+	/**
+	 * @inheritDoc
+	 */
+	public $sqlOrderBy = 'label.showOrder ASC, label.labelID ASC';
 }